Here we done with complete JWT auth manually

+++ app/services/leave_service.py
@@ -1,10 +1,10 @@
-from app.schemas.leave import Leave, EmployeeLeaveRequest
+from app.schemas.leave import Leave, EmployeeLeaveRequest, UpdateEmployeeLeave
 from app.database.leaveDB import LeaveDB
 from fastapi import HTTPException, status
 
 class LeaveService:
     def create(self, leave_id: str, leave: Leave):
-        return LeaveDB.add_Leave(leave_id, leave)
+        LeaveDB.add_Leave(leave_id, leave)
 
     def get(self, leave_id: str) -> Leave:
         leave = LeaveDB.get_Leave(leave_id)
@@ -12,10 +12,8 @@ class LeaveService:
             raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Leave not found")
         return leave
 
-    def update(self, leave_id: str, leaveUpdate: EmployeeLeaveRequest):
+    def update(self, leave_id: str, leaveUpdate: UpdateEmployeeLeave):
         existLeave = LeaveDB.get_Leave(leave_id)
-        if not existLeave:
-            raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Leave not found")
 
         if existLeave.status != "pending":
             raise HTTPException(status_code=status.HTTP_400_BAD_REQUEST, detail="Only pending leave can be updated")
@@ -24,7 +22,10 @@ class LeaveService:
         for key, value in update_data.items():
             setattr(existLeave, key, value)
 
-        return LeaveDB.update_Leave(leave_id, existLeave)
+        LeaveDB.update_Leave(leave_id, existLeave)
 
     def delete(self, leave_id: str):
-        return LeaveDB.delete_leave(leave_id)
+
+        if LeaveDB.get_Leave(leave_id).status != "pending":
+            raise HTTPException(status_code=status.HTTP_400_BAD_REQUEST, detail="Only pending leave can be deleted")
+        LeaveDB.delete_leave(leave_id)
